First working iterations of Particle and Vector (#1)
* Made Particle and Vector classes

* Added Yeast class and updated Particle generation

Added Yeast as a subclass of Particle (different radius and mass and tolerances). Generation of a particle now generates particles with varying size and mass.

* Not ignoring eclipse project files

For the sake of it being easier to u

* Adding remaining file structure folders

Added .gitignores for git to recognize them

* Unit tests for Particle, Yeast, and Vector

Made unit tests for the above classes with junit. Fixed constructor issues with Yeast/Particle. Added extra versions of add/sub methods to Vector.

* Quick number fix and format