Made Medium an interface, new Medium class and Particle class
Made a new DeionizedWater Medium and a Polystyrene1um Particle. Removed some unnecessary variable declarations in Electrode classes by putting them in the Electrode interface. Clausius-Mossotti factor computation seems to be wrong as of right now, as even when the frequency varies by orders of magnitude, the real part experiences nearly negligible changes, and never shows the behavior of switching signs as it should based on the findings in the Yafouz, Kadri, Ibrahim paper.