Updated Electrode interface and MainWindowController
Finished most of the functionality for MainWindowController for processing user input. Added a setVoltage method for the Electrode interface. Implementation TBD for each implementing class. Added a check for Simulation when trying to interrupt the threaded loop so it doesn't try to interrupt when it hasn't started yet or is already interrupted.