setVoltage(), made Application, initialized Simulation
Implemented setVoltage() for PointCharge and Dipole. Made a MIN_RADIUS variable for Electrode so charges can be calculated from desired voltages. Made the Application that launches the simulation. Initialized parts of the Simulation so that there are defaults for bounds, medium, frequency, and voltage. Moved mainWindow.fxml for ease of access.